tax_backend/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
pytesseract
openai
pillow
